Start a binary encoding document
This PR is intended to start discussion around a binary format for
module linking. This isn't supposed to be a final format nor is it even
sufficient for spec-level documentation. The hope though is that this
can be used as a basis for coordinating across the tooling ecosystem to
ensure there's a plan of record of what to implement as well as a place
for discussions.